requirements.txt
django==3.0.5
line-bot-sdk==1.18.0
whitenoise==5.2.0
gunicorn==20.0.4
django-heroku==0.3.1
dj-database-url==0.5.0
python-dotenv==0.15.0
psycopg2==2.8.6
mysqlclient==2.0.3
django-role-permissions==3.1.1
django-environ==0.8.1
django-cloudinary-storage~=0.3.0
Pillow~=8.4.0